Refactor session and origin validation for modular middleware approach
- Introduced `verifySession` and `verifyOrigin` as reusable middleware in `utils/session.js` and `utils/origin.js`.
- Refactored `auth.js`, `generate-image.js`, and `translate.js` to use `onRequestPost` with modular validation steps:
  - `verifySession`: Validates session cookies for authenticated access.
  - `verifyOrigin`: Validates the request's `Origin` against allowed origins.
- Replaced inline session and origin validation logic with centralized utilities for improved maintainability.
- Renamed `translatePrompt` to `translate` for consistency in `translate.js`.
- Updated `utils/session.js` to include a `verifySession` middleware function for streamlined session validation.
- Added a new file `utils/origin.js` to handle origin validation logic.
- Ensured all middleware functions return appropriate `Response` objects for error handling or call `context.next()` to continue execution.
