+++ functions/auth.js
@@ -1,13 +1,16 @@
+import { verifyOrigin } from "../utils/origin.js";
 import turnstilePlugin from "@cloudflare/pages-plugin-turnstile";
 import { createSignedCookie } from "../utils/session";
 
-// 1. Turnstile認証とセッションCookieの払い出しAPI
 export const onRequestPost = [
+  verifyOrigin,
+  // Turnstile認証
   async (context) => {
     return turnstilePlugin({
       secret: context.env.TRUNSTILE_SECRET_KEY,
     })(context);
   },
+  // セッションCookieの払い出し
   async ({ request, env }) => {
     const sessionSecret = new TextEncoder().encode(env.SESSION_SECRET);
     const response = new Response("Authenticated", { status: 200 });
