+++ functions/generate-image.js
@@ -1,26 +1,9 @@
-import { validateSessionCookie } from "../utils/session.js";
+import { verifySession } from "../utils/session.js";
+import { verifyOrigin } from "../utils/origin.js";
 
-export async function onRequestPost({ request, env }) {
-  const sessionSecret = new TextEncoder().encode(env.SESSION_SECRET);
+export const onRequestPost = [verifySession, verifyOrigin, handleGenerateImage];
 
-  // セッションCookieの検証
-  const isValidSession = await validateSessionCookie(request, sessionSecret);
-  if (!isValidSession) {
-    console.error("認証エラー: 無効なセッションCookie");
-    return new Response("Unauthorized", { status: 401 });
-  }
-
-  // Origin検証
-  const allowedOrigin = env.ALLOWED_ORIGIN;
-  const origin = request.headers.get("Origin");
-  if (allowedOrigin !== origin) {
-    console.error(
-      `許可されていないOriginです。origin=${origin}, allowedOrigin=${allowedOrigin}`
-    );
-    return new Response("Forbidden", { status: 403 });
-  }
-
-  // 画像生成の処理
+async function handleGenerateImage({ request, env }) {
   const formData = await request.formData();
   const prompt = formData.get("prompt");
   return await generateImage(prompt, env);
